Add auth helper case
