+++ lib/elixir_google_scraper_web/controllers/api/v1/keyword_controller.ex
@@ -2,8 +2,7 @@ defmodule ElixirGoogleScraperWeb.Api.V1.KeywordController do
   use ElixirGoogleScraperWeb, :controller
 
   alias ElixirGoogleScraper.Scraper.Keywords
-  alias ElixirGoogleScraperWeb.V1.ErrorView
-  alias ElixirGoogleScraperWeb.V1.KeywordView
+  alias ElixirGoogleScraperWeb.V1.{ErrorView, KeywordView}
 
   def index(conn, params) do
     {keywords, pagination} = Keywords.paginated_user_keywords(conn.assigns.current_user, params)
